requirements.txt
# Core Dependencies to install 
# Use --> pip install -r requirements.txt
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0